refactor(work-brief): split monolith into phase-registry dispatcher pattern
Extract 733-line brief-next.js into thin orchestrator (~313 lines) backed by
per-phase modules (inputs, overlap, draft, validate, memorize, done). Match
dispatcher architecture used by check2. Add phase-registry.test.js with 9
dispatcher tests; all 21 brief tests pass.
